+++ javascript/node/selenium-webdriver/lib/test/index.js
@@ -28,6 +28,9 @@ var build = require('./build'),
     fileserver = require('./fileserver');
 
 
+const LEGACY_FIREFOX = 'legacy-' + webdriver.Browser.FIREFOX;
+
+
 /**
  * Browsers with native support.
  * @type {!Array.<webdriver.Browser>}
@@ -36,6 +39,7 @@ var NATIVE_BROWSERS = [
   webdriver.Browser.CHROME,
   webdriver.Browser.EDGE,
   webdriver.Browser.FIREFOX,
+  LEGACY_FIREFOX,
   webdriver.Browser.IE,
   webdriver.Browser.OPERA,
   webdriver.Browser.PHANTOM_JS,
@@ -46,9 +50,9 @@ var NATIVE_BROWSERS = [
 var serverJar = process.env['SELENIUM_SERVER_JAR'];
 var remoteUrl = process.env['SELENIUM_REMOTE_URL'];
 var useLoopback = process.env['SELENIUM_USE_LOOP_BACK'] == '1';
+var noMarionette = /^0|false$/i.test(process.env['SELENIUM_GECKODRIVER']);
 var startServer = !!serverJar && !remoteUrl;
 var nativeRun = !serverJar && !remoteUrl;
-
 
 var browsersToTest = (function() {
   var permitRemoteBrowsers = !!remoteUrl || !!serverJar;
@@ -63,12 +67,20 @@ var browsersToTest = (function() {
     if (parts[0] === 'edge') {
       parts[0] = webdriver.Browser.EDGE;
     }
+    if (noMarionette && parts[0] === webdriver.Browser.FIREFOX) {
+      parts[0] = LEGACY_FIREFOX;
+    }
     return parts.join(':');
   });
+
   browsers.forEach(function(browser) {
     var parts = browser.split(/:/, 3);
     if (parts[0] === 'ie') {
       parts[0] = webdriver.Browser.IE;
+    }
+
+    if (parts[0] === LEGACY_FIREFOX) {
+      return;
     }
 
     if (NATIVE_BROWSERS.indexOf(parts[0]) == -1 && !permitRemoteBrowsers) {
@@ -145,6 +157,11 @@ function TestEnvironment(browserName, server) {
 
     builder.build = function() {
       var parts = browserName.split(/:/, 3);
+
+      if (parts[0] === LEGACY_FIREFOX) {
+        parts[0] = webdriver.Browser.FIREFOX;
+      }
+
       builder.forBrowser(parts[0], parts[1], parts[2]);
       if (server) {
         builder.usingServer(server.address());
@@ -199,7 +216,7 @@ function suite(fn, opt_options) {
       testing.describe('[' + browser + ']', function() {
 
         if (isDevMode && nativeRun) {
-          if (browser === webdriver.Browser.FIREFOX) {
+          if (browser === LEGACY_FIREFOX) {
             testing.before(function() {
               return build.of('//javascript/firefox-driver:webdriver')
                   .onlyOnce().go();