+++ javascript/node/selenium-webdriver/CHANGES.md
@@ -9,6 +9,9 @@
 * Internal cleanup replacing uses of managed promises with native promises
 * Removed the mandatory use of Firefox Dev Edition, when using Marionette driver
 * Fixed timeouts' URL
+
+### API Changes
+
 * `promise.Deferred` is no longer a thenable object.
 * `Options#addCookie()` now takes a record object instead of 7 individual
   parameters. A TypeError will be thrown if addCookie() is called with invalid
@@ -16,6 +19,7 @@
 * When adding cookies, the desired expiry must be provided as a Date or in
   _seconds_ since epoch. When retrieving cookies, the expiration is always
   returned in seconds.
+*  Renamed `firefox.Options#useMarionette` to `firefox.Options#useGeckoDriver`
 * Removed deprecated modules:
    - `selenium-webdriver/error` (use `selenium-webdriver/lib/error`,\
      or the `error` property exported by `selenium-webdriver`)